Lost and Found Items
Introduction of project 
Introducing the Lost and Found Management System – a modern solution designed to transform the way we manage lost and found items. We understand the emotional and practical challenges associated with losing possessions and the desire to do the right thing when finding someone else's belongings. Our system is here to simplify and streamline this process, making it easier and more efficient for everyone involved. 

Scope of the Project The scope of the Lost and Found Management System project encompasses a range of functionalities and objectives designed to facilitate the efficient tracking, reporting, and recovery of lost items. This system provides a comprehensive solution to address the challenges associated with lost and found items within various contexts, including public spaces, transportation, educational institutions, and beyond.